+++ changelog.md
@@ -7,6 +7,7 @@
     - Breaking: add missing `Metadata` constructors: `AlbumSort`, `OriginalDate`, `Conductor`, `Work`, `Grouping`, `Label`, `MUSICBRAINZ_WORKID`.
     - Breaking: add missing `Subsystem` constructors: `PartitionS`, `NeighborS`, `MountS`.
     - Breaking: add missing `ReplayGainMode` constructor: `AutoMode`.
+    - Breaking: move `toggle` from `N.M.C.Extensions` to `N.M.(C|A).PlaybackControl`.
 
 * v0.9.4.0 unreleased
     - Add monadic versions of `searchAdd` and `searchAddPl` commands (previously